Refactor jdbc chat memory repository autoconfigurations and Spring Boot starters
 - Rename the existing chat memory repository autoconfigurations to include the suffix "repository" for clarity

   spring-ai-autoconfigure-model-chat-memory-cassandra -> spring-ai-autoconfigure-model-chat-memory-repository-cassandra
   spring-ai-autoconfigure-model-chat-memory-jdbc -> spring-ai-autoconfigure-model-chat-memory-repository-jdbc
   spring-ai-autoconfigure-model-chat-memory-neo4j -> spring-ai-autoconfigure-model-chat-memory-repository-neo4j

 - Rename the packages of the classes to include "repository"
 - Rename the classes to have "ChatMemoryRepository"
 - Update the starters to include the renaming
 - Update Spring AI BOM and parent POM files
