+++ spring-ai-bom/pom.xml
@@ -481,21 +481,23 @@
 				<version>${project.version}</version>
 			</dependency>
 
+			<!-- Spring AI Chat memory Repository auto-configurations -->
+
 			<dependency>
 				<groupId>org.springframework.ai</groupId>
-				<artifactId>spring-ai-autoconfigure-model-chat-memory-cassandra</artifactId>
+				<artifactId>spring-ai-autoconfigure-model-chat-memory-repository-cassandra</artifactId>
 				<version>${project.version}</version>
 			</dependency>
 
 			<dependency>
 				<groupId>org.springframework.ai</groupId>
-				<artifactId>spring-ai-autoconfigure-model-chat-memory-jdbc</artifactId>
+				<artifactId>spring-ai-autoconfigure-model-chat-memory-repository-jdbc</artifactId>
 				<version>${project.version}</version>
 			</dependency>
 
 			<dependency>
 				<groupId>org.springframework.ai</groupId>
-				<artifactId>spring-ai-autoconfigure-model-chat-memory-neo4j</artifactId>
+				<artifactId>spring-ai-autoconfigure-model-chat-memory-repository-neo4j</artifactId>
 				<version>${project.version}</version>
 			</dependency>
 
@@ -1017,23 +1019,23 @@
 				<version>${project.version}</version>
 			</dependency>
 
-			<!-- Spring AI Spring Boot starters for Chat Memory -->
+			<!-- Spring AI Spring Boot starters for Chat Memory Repository -->
 
 			<dependency>
 				<groupId>org.springframework.ai</groupId>
-				<artifactId>spring-ai-starter-model-chat-memory-cassandra</artifactId>
+				<artifactId>spring-ai-starter-model-chat-memory-repository-cassandra</artifactId>
 				<version>${project.version}</version>
 			</dependency>
 
 			<dependency>
 				<groupId>org.springframework.ai</groupId>
-				<artifactId>spring-ai-starter-model-chat-memory-jdbc</artifactId>
+				<artifactId>spring-ai-starter-model-chat-memory-repository-jdbc</artifactId>
 				<version>${project.version}</version>
 			</dependency>
 
 			<dependency>
 				<groupId>org.springframework.ai</groupId>
-				<artifactId>spring-ai-starter-model-chat-memory-neo4j</artifactId>
+				<artifactId>spring-ai-starter-model-chat-memory-repository-neo4j</artifactId>
 				<version>${project.version}</version>
 			</dependency>
